+++ HISTORY.rst
@@ -7,3 +7,4 @@ History
 
 * Implements the Choreographer Worker (ericof)
 * Mail action, now, accepts sender email and sender name (ericof)
+* BODY-95: Implement forgot password actions (ericof)